web/requirements.txt
Django>=3.2,<5.0
djangorestframework>=3.14
PyYAML>=6.0
